+++ cpp/bootlang/include/nodes/break_node.hpp
@@ -4,7 +4,7 @@
 
 class BreakNode : public Node {
 public:
-    BreakNode(uint32_t lineno, uint32_t col);
+    BreakNode(size_t lineno, size_t col);
 
     std::string toCode(int indent) const override;
 };
