+++ cpp/bootlang/include/nodes/dictliteral_node.hpp
@@ -9,7 +9,7 @@ class DictLiteralNode : public Node {
 public:
     std::vector<DictPair> elems;
 
-    DictLiteralNode(uint32_t lineno, uint32_t col, std::vector<DictPair>&& elems);
+    DictLiteralNode(size_t lineno, size_t col, std::vector<DictPair>&& elems);
 
     std::string toCode(int indent) const override;
 };
