+++ cpp/bootlang/include/nodes/if_node.hpp
@@ -13,7 +13,7 @@ public:
     std::vector<CondBlock> elif_blocks;
     std::optional<BlockNodePtr> else_block;
 
-    IfNode(uint32_t lineno, uint32_t col, CondBlock&& if_block, std::vector<CondBlock>&& elif_blocks, std::optional<BlockNodePtr>&& else_block);
+    IfNode(size_t lineno, size_t col, CondBlock&& if_block, std::vector<CondBlock>&& elif_blocks, std::optional<BlockNodePtr>&& else_block);
 
     std::string toCode(int indent) const override;
 };
