+++ cpp/bootlang/include/nodes/index_node.hpp
@@ -7,7 +7,7 @@ public:
     NodePtr left;
     NodePtr right;
 
-    IndexNode(uint32_t lineno, uint32_t col, NodePtr&& left, NodePtr&& right);
+    IndexNode(size_t lineno, size_t col, NodePtr&& left, NodePtr&& right);
 
     std::string toCode(int indent) const override;
 };
