+++ cpp/bootlang/include/nodes/int_node.hpp
@@ -6,7 +6,7 @@ class IntNode : public Node {
 public:
     long value;
 
-    IntNode(uint32_t lineno, uint32_t col, long value);
+    IntNode(size_t lineno, size_t col, long value);
 
     std::string toCode(int indent) const override;
 };
