+++ cpp/bootlang/include/nodes/listliteral_node.hpp
@@ -7,7 +7,7 @@ class ListLiteralNode : public Node {
 public:
     std::vector<NodePtr> elems;
 
-    ListLiteralNode(uint32_t lineno, uint32_t col, std::vector<NodePtr>&& elems);
+    ListLiteralNode(size_t lineno, size_t col, std::vector<NodePtr>&& elems);
 
     std::string toCode(int indent) const override;
 };
