+++ cpp/bootlang/include/nodes/propertyaccess_node.hpp
@@ -8,7 +8,7 @@ public:
     NodePtr left;
     std::string property;
 
-    PropertyAccessNode(uint32_t lineno, uint32_t col, NodePtr&& left, std::string property);
+    PropertyAccessNode(size_t lineno, size_t col, NodePtr&& left, std::string property);
 
     std::string toCode(int indent) const override;
 };
