+++ cpp/bootlang/include/nodes/return_node.hpp
@@ -7,7 +7,7 @@ class ReturnNode : public Node {
 public:
     std::optional<NodePtr> value;
 
-    ReturnNode(uint32_t lineno, uint32_t col, std::optional<NodePtr>&& value);
+    ReturnNode(size_t lineno, size_t col, std::optional<NodePtr>&& value);
 
     std::string toCode(int indent) const override;
 };
