+++ cpp/bootlang/include/nodes/string_node.hpp
@@ -7,7 +7,7 @@ class StringNode : public Node {
 public:
     std::string value;
 
-    StringNode(uint32_t lineno, uint32_t col, std::string value);
+    StringNode(size_t lineno, size_t col, std::string value);
 
     std::string toCode(int indent) const override;
 };
