+++ cpp/bootlang/include/nodes/tupleliteral_node.hpp
@@ -7,7 +7,7 @@ class TupleLiteralNode : public Node {
 public:
     std::vector<NodePtr> elems;
 
-    TupleLiteralNode(uint32_t lineno, uint32_t col, std::vector<NodePtr>&& elems);
+    TupleLiteralNode(size_t lineno, size_t col, std::vector<NodePtr>&& elems);
 
     std::string toCode(int indent) const override;
 };
