+++ cpp/bootlang/include/nodes/unaryop_node.hpp
@@ -8,7 +8,7 @@ public:
     NodePtr right;
     Token::Type op;
 
-    UnaryOpNode(uint32_t lineno, uint32_t col, NodePtr&& right, Token::Type op);
+    UnaryOpNode(size_t lineno, size_t col, NodePtr&& right, Token::Type op);
 
     std::string toCode(int indent) const override;
 };
