+++ cpp/bootlang/include/nodes/while_node.hpp
@@ -9,7 +9,7 @@ public:
     NodePtr cond;
     BlockNodePtr block;
 
-    WhileNode(uint32_t lineno, uint32_t col, NodePtr&& cond, BlockNodePtr&& block);
+    WhileNode(size_t lineno, size_t col, NodePtr&& cond, BlockNodePtr&& block);
 
     std::string toCode(int indent) const override;
 };
