+++ cpp/bootlang/include/nodes/yield_node.hpp
@@ -7,7 +7,7 @@ class YieldNode : public Node {
   public:
     std::optional<NodePtr> value;
 
-    YieldNode(uint32_t lineno, uint32_t col, std::optional<NodePtr>&& value);
+    YieldNode(size_t lineno, size_t col, std::optional<NodePtr>&& value);
 
     std::string toCode(int indent) const override;
 };
