+++ cpp/bootlang/src/nodes/binaryop_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/binaryop_node.hpp"
 
-BinaryOpNode::BinaryOpNode(uint32_t lineno, uint32_t col, NodePtr&& left, NodePtr&& right, Token::Type op)
+BinaryOpNode::BinaryOpNode(size_t lineno, size_t col, NodePtr&& left, NodePtr&& right, Token::Type op)
     : Node(lineno, col, Node::Type::BINARY_OP), left(std::move(left)), right(std::move(right)), op(op) {}
 
 std::string BinaryOpNode::toCode(int indent) const {
