+++ cpp/bootlang/src/nodes/block_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/block_node.hpp"
 
-BlockNode::BlockNode(uint32_t lineno, uint32_t col, std::vector<NodePtr>&& stmts)
+BlockNode::BlockNode(size_t lineno, size_t col, std::vector<NodePtr>&& stmts)
     : Node(lineno, col, Node::Type::BLOCK), stmts(std::move(stmts)) {}
 
 std::string BlockNode::toCode(int indent) const {
