+++ cpp/bootlang/src/nodes/bool_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/bool_node.hpp"
 
-BoolNode::BoolNode(uint32_t lineno, uint32_t col, bool value)
+BoolNode::BoolNode(size_t lineno, size_t col, bool value)
     : Node(lineno, col, Node::Type::BOOL), value(value) {}
 
 std::string BoolNode::toCode(int indent) const {
