+++ cpp/bootlang/src/nodes/break_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/break_node.hpp"
 
-BreakNode::BreakNode(uint32_t lineno, uint32_t col)
+BreakNode::BreakNode(size_t lineno, size_t col)
     : Node(lineno, col, Node::Type::BREAK) {}
 
 std::string BreakNode::toCode(int indent) const {
