+++ cpp/bootlang/src/nodes/call_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/call_node.hpp"
 
-CallNode::CallNode(uint32_t lineno, uint32_t col, NodePtr&& left, std::vector<NodePtr>&& args)
+CallNode::CallNode(size_t lineno, size_t col, NodePtr&& left, std::vector<NodePtr>&& args)
     : Node(lineno, col, Node::Type::CALL), left(std::move(left)), args(std::move(args)) {}
 
 std::string CallNode::toCode(int indent) const {
