+++ cpp/bootlang/src/nodes/continue.node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/continue_node.hpp"
 
-ContinueNode::ContinueNode(uint32_t lineno, uint32_t col)
+ContinueNode::ContinueNode(size_t lineno, size_t col)
     : Node(lineno, col, Node::Type::CONTINUE) {}
 
 std::string ContinueNode::toCode(int indent) const {
