+++ cpp/bootlang/src/nodes/dictliteral_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/dictliteral_node.hpp"
 
-DictLiteralNode::DictLiteralNode(uint32_t lineno, uint32_t col, std::vector<DictPair>&& elems)
+DictLiteralNode::DictLiteralNode(size_t lineno, size_t col, std::vector<DictPair>&& elems)
 : Node(lineno, col, Node::Type::DICT_LITERAL), elems(std::move(elems)) {}
 
 std::string DictLiteralNode::toCode(int indent) const {
