+++ cpp/bootlang/src/nodes/float_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/float_node.hpp"
 
-FloatNode::FloatNode(uint32_t lineno, uint32_t col, double value)
+FloatNode::FloatNode(size_t lineno, size_t col, double value)
     : Node(lineno, col, Node::Type::FLOAT), value(value) {}
 
 std::string FloatNode::toCode(int indent) const {
