+++ cpp/bootlang/src/nodes/for_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/for_node.hpp"
 
-ForNode::ForNode(uint32_t lineno, uint32_t col, NodePtr&& target, NodePtr&& iterable, BlockNodePtr&& block)
+ForNode::ForNode(size_t lineno, size_t col, NodePtr&& target, NodePtr&& iterable, BlockNodePtr&& block)
 : Node(lineno, col, Node::Type::FOR), target(std::move(target)), iterable(std::move(iterable)), block(std::move(block)) {}
 
 std::string ForNode::toCode(int indent) const {
