+++ cpp/bootlang/src/nodes/index_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/index_node.hpp"
 
-IndexNode::IndexNode(uint32_t lineno, uint32_t col, NodePtr&& left, NodePtr&& right) 
+IndexNode::IndexNode(size_t lineno, size_t col, NodePtr&& left, NodePtr&& right) 
 : Node(lineno, col, Node::Type::INDEX), left(std::move(left)), right(std::move(right)) {}
 
 std::string IndexNode::toCode(int indent) const {
