+++ cpp/bootlang/src/nodes/setliteral_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/setliteral_node.hpp"
 
-SetLiteralNode::SetLiteralNode(uint32_t lineno, uint32_t col, std::vector<NodePtr>&& elems)
+SetLiteralNode::SetLiteralNode(size_t lineno, size_t col, std::vector<NodePtr>&& elems)
 : Node(lineno, col, Node::Type::SET_LITERAL), elems(std::move(elems)) {}
 
 std::string SetLiteralNode::toCode(int indent) const {
