+++ cpp/bootlang/src/nodes/string_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/string_node.hpp"
 
-StringNode::StringNode(uint32_t lineno, uint32_t col, std::string value)
+StringNode::StringNode(size_t lineno, size_t col, std::string value)
     : Node(lineno, col, Node::Type::STRING), value(value) {}
 
 std::string StringNode::toCode(int indent) const {
