+++ cpp/bootlang/src/nodes/ternary_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/ternary_node.hpp"
 
-TernaryNode::TernaryNode(uint32_t lineno, uint32_t col, NodePtr&& left, NodePtr&& right, NodePtr&& cond)
+TernaryNode::TernaryNode(size_t lineno, size_t col, NodePtr&& left, NodePtr&& right, NodePtr&& cond)
     : Node(lineno, col, Node::Type::TERNARY), left(std::move(left)), right(std::move(right)), cond(std::move(cond)) {}
 
 std::string TernaryNode::toCode(int indent) const {
