+++ cpp/bootlang/src/nodes/tupleliteral_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/tupleliteral_node.hpp"
 
-TupleLiteralNode::TupleLiteralNode(uint32_t lineno, uint32_t col, std::vector<NodePtr>&& elems)
+TupleLiteralNode::TupleLiteralNode(size_t lineno, size_t col, std::vector<NodePtr>&& elems)
 : Node(lineno, col, Node::Type::TUPLE_LITERAL), elems(std::move(elems)) {}
 
 std::string TupleLiteralNode::toCode(int indent) const {
