+++ cpp/bootlang/src/nodes/while_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/while_node.hpp"
 
-WhileNode::WhileNode(uint32_t lineno, uint32_t col, NodePtr&& cond, BlockNodePtr&& block)
+WhileNode::WhileNode(size_t lineno, size_t col, NodePtr&& cond, BlockNodePtr&& block)
 : Node(lineno, col, Node::Type::WHILE), cond(std::move(cond)), block(std::move(block)) {}
 
 std::string WhileNode::toCode(int indent) const {
