+++ cpp/bootlang/src/nodes/yield_node.cpp
@@ -1,6 +1,6 @@
 #include "nodes/yield_node.hpp"
 
-YieldNode::YieldNode(uint32_t lineno, uint32_t col, std::optional<NodePtr>&& value)
+YieldNode::YieldNode(size_t lineno, size_t col, std::optional<NodePtr>&& value)
     : Node(lineno, col, Node::Type::YIELD), value(std::move(value)) {}
 
 std::string YieldNode::toCode(int indent) const {
